requirements/base.txt
Django==1.11.8
python-vlc
psycopg2
Pillow
Unipath
django-multiselectfield